split levelset export concerns off into separate script
